Rename to "System.Linq.Dynamic.Core" and add more frameworks.
