--- Src/System.Linq.Dynamic.ConsoleTestApp.net40/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentValidation.dnx" version="2.0.0-beta002" targetFramework="net40" />
-  <package id="ReflectionBridge" version="0.0.6" targetFramework="net40" />
-</packages>
