+++ Src/System.Linq.Dynamic.Core.ConsoleTestApp.net40/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="JetBrains.Annotations" version="10.0.0" targetFramework="net40" />
+  <package id="ReflectionBridge" version="0.0.8" targetFramework="net40" />
+</packages>
