+++ Src/System.Linq.Dynamic.Core.Tests/Helpers/Entities/Post.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace System.Linq.Dynamic.Tests.Helpers.Entities
+namespace System.Linq.Dynamic.Core.Tests.Helpers.Entities
 {
     public class Post
     {
