+++ Src/System.Linq.Dynamic.Core.Tests/System.Linq.Dynamic.Core.Tests.xproj
@@ -8,7 +8,7 @@
   <Import Project="$(VSToolsPath)\DNX\Microsoft.DNX.targets" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>45d205d1-2338-438a-a850-52a9677ec924</ProjectGuid>
-    <RootNamespace>System.Linq.Dynamic.Tests</RootNamespace>
+    <RootNamespace>System.Linq.Dynamic.Core.Tests</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">..\..\artifacts\obj\$(MSBuildProjectName)</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">..\..\artifacts\bin\$(MSBuildProjectName)\</OutputPath>
   </PropertyGroup>
