+++ Src/System.Linq.Dynamic.Core.Tests/project.json
@@ -2,11 +2,10 @@
     "version": "1.0.0-*",
 
     "dependencies": {
-        "xunit": "2.1.0",
-        "xunit.runner.dnx": "2.1.0-rc1-build204",
+        "System.Linq.Dynamic.Core": "1.0.0-*",
 
-        "TestToolsToXunitProxy": "0.0.1",
-        "System.Linq.Dynamic.Core": "2.0.0-*"
+        "xunit": "2.1.0",
+        "TestToolsToXunitProxy": "0.0.1"
     },
 
     "commands": {
@@ -16,6 +15,8 @@
     "frameworks": {
         "dnx452": {
             "dependencies": {
+                "xunit.runner.dnx": "2.1.0-rc1-build204",
+
                 "EntityFramework.Core": "7.0.0-rc1-final",
                 "EntityFramework.Sqlite": "7.0.0-rc1-final",
                 "EntityFramework.InMemory": "7.0.0-rc1-final",
@@ -27,6 +28,8 @@
 
         "dnxcore50": {
             "dependencies": {
+                "xunit.runner.dnx": "2.1.0-rc1-build204",
+
                 "EntityFramework.Core": "7.0.0-rc1-final",
                 "EntityFramework.Sqlite": "7.0.0-rc1-final",
                 "EntityFramework.InMemory": "7.0.0-rc1-final",
