+++ Src/System.Linq.Dynamic.Core.sln
@@ -28,6 +28,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "web", "web", "{7F147C37-B30
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamicLinqWebDocs", "DynamicLinqWebDocs\DynamicLinqWebDocs.csproj", "{1942BDB5-CDA0-4B50-85BD-D726303F219C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src-console", "src-console", "{C98BB543-9BDA-4EE3-927D-0F10CC24A49B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "System.Linq.Dynamic.Core.ConsoleTestApp.net40", "System.Linq.Dynamic.Core.ConsoleTestApp.net40\System.Linq.Dynamic.Core.ConsoleTestApp.net40.csproj", "{0EB12661-F5CF-4071-9812-F4C8AF3D4531}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CE1620BD-A313-4223-B991-467DA346B2ED} = {CE1620BD-A313-4223-B991-467DA346B2ED}
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -58,16 +65,21 @@ Global
 		{1942BDB5-CDA0-4B50-85BD-D726303F219C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1942BDB5-CDA0-4B50-85BD-D726303F219C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1942BDB5-CDA0-4B50-85BD-D726303F219C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0EB12661-F5CF-4071-9812-F4C8AF3D4531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0EB12661-F5CF-4071-9812-F4C8AF3D4531}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0EB12661-F5CF-4071-9812-F4C8AF3D4531}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0EB12661-F5CF-4071-9812-F4C8AF3D4531}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{CE1620BD-A313-4223-B991-467DA346B2ED} = {DB9A5C42-3F8B-4546-96F2-DD35BCACA018}
-		{E1B36B40-0134-46C4-922B-7EC1FF6CD90F} = {DB9A5C42-3F8B-4546-96F2-DD35BCACA018}
+		{E1B36B40-0134-46C4-922B-7EC1FF6CD90F} = {C98BB543-9BDA-4EE3-927D-0F10CC24A49B}
 		{45D205D1-2338-438A-A850-52A9677EC924} = {32699374-E6B9-4870-9D4A-C8884D16CEA9}
 		{58BDDB6A-B81E-4F7D-877B-BB23DA5E5A24} = {5FB356F8-05AB-4CA5-A43C-7CA5AD943166}
 		{F3DC7D8A-1A63-4CE0-AEA5-A83EC5CDD873} = {5FB356F8-05AB-4CA5-A43C-7CA5AD943166}
 		{1942BDB5-CDA0-4B50-85BD-D726303F219C} = {7F147C37-B300-4C82-B98A-DF25CEA149E2}
+		{0EB12661-F5CF-4071-9812-F4C8AF3D4531} = {C98BB543-9BDA-4EE3-927D-0F10CC24A49B}
 	EndGlobalSection
 EndGlobal
