+++ Src/System.Linq.Dynamic.Core/DynamicExpression.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq.Expressions;
 
-namespace System.Linq.Dynamic
+namespace System.Linq.Dynamic.Core
 {
     internal static class DynamicExpression
     {
