+++ Src/System.Linq.Dynamic.Core/DynamicObjectClass.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Dynamic;
 using System.Text;
 
-namespace System.Linq.Dynamic
+namespace System.Linq.Dynamic.Core
 {
     /// <summary>
     /// A DynamicClass as Replacement for creating classes via Reflection Emit, wich is not supported in WinRT
