+++ Src/System.Linq.Dynamic.Core/DynamicOrdering.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq.Expressions;
 
-namespace System.Linq.Dynamic
+namespace System.Linq.Dynamic.Core
 {
     internal class DynamicOrdering
     {
