+++ Src/System.Linq.Dynamic.Core/ExpressionParser.NetFX_backup.cs
@@ -5,7 +5,7 @@ using System.Reflection;
 using System.Collections;
 using ReflectionBridge.Extensions;
 
-namespace System.Linq.Dynamic
+namespace System.Linq.Dynamic.Core
 {
     internal class ExpressionParser2
     {
